Create a general function utils module
